Change to folder struckture
