+++ vite-frontend/src/App.tsx
@@ -1,7 +1,7 @@
 //import "./App.css";
 import { EmailClient } from "./containers/EmailClient";
 import Navbar from "./components/Navbar";
-import Hero from "./components/Hero";
+import Hero from "./components/Hero/Hero";
 
 function App() {
   return (
