+++ vite-frontend/src/components/Hero/ImageCircle.tsx
